backend/pyproject.toml
[project]
name = "postmortem-backend"
version = "0.1.0"
description = "Postmortem backend — FastAPI service for decision archaeology"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn[standard]>=0.38.0",
    "sse-starlette>=3.0.2",
    "httpx>=0.28.1",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.11.0",
    "anthropic>=0.73.0",
    "python-dotenv>=1.1.1",
    "duckdb>=1.4.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
    "black>=25.11.0",
    "mypy>=1.16.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.1",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC", "RUF"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["sse_starlette.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
